HAPP-1467 - updated patient v2 api response object.
