+++ data/src/main/java/ca/bc/gov/data/datasource/remote/PatientRemoteDataSource.kt
@@ -4,7 +4,7 @@ import ca.bc.gov.common.const.MESSAGE_INVALID_RESPONSE
 import ca.bc.gov.common.const.SERVER_ERROR
 import ca.bc.gov.common.exceptions.MyHealthException
 import ca.bc.gov.data.datasource.remote.api.HealthGatewayPrivateApi
-import ca.bc.gov.data.datasource.remote.model.base.patient.PatientPayload
+import ca.bc.gov.data.datasource.remote.model.response.PatientResponse
 import ca.bc.gov.data.utils.safeCall
 import javax.inject.Inject
 
@@ -12,10 +12,8 @@ class PatientRemoteDataSource @Inject constructor(
     private val healthGatewayPrivateApi: HealthGatewayPrivateApi
 ) {
 
-    suspend fun getPatient(token: String, hdid: String): PatientPayload {
-        val response = safeCall { healthGatewayPrivateApi.getPatient(token = token, hdid = hdid) }
+    suspend fun getPatient(token: String, hdid: String): PatientResponse {
+        return safeCall { healthGatewayPrivateApi.getPatient(token = token, hdid = hdid) }
             ?: throw MyHealthException(SERVER_ERROR, MESSAGE_INVALID_RESPONSE)
-
-        return response.payload ?: throw MyHealthException(SERVER_ERROR, MESSAGE_INVALID_RESPONSE)
     }
 }
